% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-functions.dust.R
\name{annotate.by.cdd.region}
\alias{annotate.by.cdd.region}
\title{Annotate mutations by CDD region. It requires the output.folder contains a file from running the parse.aggregated.mut() function. That is, prefix.mut.filtered.txt file must exist.}
\usage{
annotate.by.cdd.region(output.folder = "", output.prefix = "")
}
\arguments{
\item{output.folder:}{the folder to write output files}

\item{output.prefix:}{prefix used to name the output file as prefix.mut.region.txt.}
}
\description{
Annotate mutations by CDD region. It requires the output.folder contains a file from running the parse.aggregated.mut() function. That is, prefix.mut.filtered.txt file must exist.
}
\examples{
annotate.by.cdd.region(output.folder='./examples/', output.prefix='TCGA.CHOL')
}
