% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_aux_functions.R
\name{backtracking_changepoint}
\alias{backtracking_changepoint}
\title{backtracking_changepoint}
\usage{
backtracking_changepoint(cp, n)
}
\arguments{
\item{cp}{vector of changepoints of size n+1}

\item{n}{data length}
}
\value{
An increasing integer vector of change-point positions, ending at \code{n}.
}
\description{
Internal helper that walks the \code{cp} back-pointer vector
  produced by a segmentation algorithm to recover the ordered sequence of
  change points. Not meant to be called directly on raw time series data.
}
\keyword{internal}
