% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_plot_dual.R
\name{barplot_dual_1D}
\alias{barplot_dual_1D}
\title{barplot_dual_1D}
\usage{
barplot_dual_1D(
  nb = 1000,
  s1 = 18,
  s2 = 15,
  n = 20,
  oneParam = 0,
  type = "gauss",
  penalty = 2 * log(n),
  pruningOpt = 3
)
}
\arguments{
\item{nb}{number of simulations}

\item{s1}{index for q_t^s1}

\item{s2}{index for q_t^s2}

\item{n}{data length}

\item{oneParam}{value of the unique parameter of the unique segment}

\item{type}{type of cost to use: \code{"gauss"}, \code{"exp"}, \code{"poisson"}, \code{"geom"}, \code{"bern"}, \code{"binom"}, \code{"negbin"}}

\item{penalty}{penalty value (non-negative)}

\item{pruningOpt}{pruningOpt pruning option for \code{dust_R_1D}.}
}
\value{
A barplot counting the number of time at each mu fixed, the dual value evaluated at mu exceeds the pruning threshold. The value returned by the function is the number of time s1 has not been pruned by s2. In case of pruning, we don't consider the assiciated simulation in the barplot.
}
\description{
identifying the mu values for which we prune with nb repetitions of the simulation
}
\examples{
barplot_dual_1D()
}
