% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-functions.dust.R
\name{calculate.expected.by.region}
\alias{calculate.expected.by.region}
\title{Calculate the expected frequencies for CDD regions. It requires the output.folder contains a file from running the annotate.by.cdd.region() function. That is, prefix.mut.region.txt file must exist.}
\usage{
calculate.expected.by.region(
  output.folder = output.folder,
  output.prefix = output.prefix
)
}
\arguments{
\item{output.folder:}{the folder to write output files}

\item{output.prefix:}{prefix used to name the output file as prefix.exp_seq.by_type.region.txt.}
}
\description{
Calculate the expected frequencies for CDD regions. It requires the output.folder contains a file from running the annotate.by.cdd.region() function. That is, prefix.mut.region.txt file must exist.
}
\examples{
calculate.expected.by.region(output.folder='./examples/', output.prefix='TCGA.CHOL')
}
