% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-functions.dust.R
\name{compute.selection.by.region}
\alias{compute.selection.by.region}
\title{Estimate selection coefficients of missense mutations and truncating mutations for CDD regions. It requires the output.folder contains files from running the parse.aggregated.mut(), annotate.by.cdd.region(), and calculate.expected.by.region() functions. That is, prefix.mut.filtered.txt, prefix.mut.region.txt, and prefix.exp_seq.by_type.region.txt files must exist.}
\usage{
compute.selection.by.region(
  output.folder = output.folder,
  output.prefix = output.prefix,
  N.sim = 1000
)
}
\arguments{
\item{output.folder:}{the folder to write output files}

\item{output.prefix:}{prefix used to name the output file as prefix.selection.region.txt.}

\item{N.sim:}{the number of simulations to perform to estimate false positives (default: 1000).}
}
\description{
Estimate selection coefficients of missense mutations and truncating mutations for CDD regions. It requires the output.folder contains files from running the parse.aggregated.mut(), annotate.by.cdd.region(), and calculate.expected.by.region() functions. That is, prefix.mut.filtered.txt, prefix.mut.region.txt, and prefix.exp_seq.by_type.region.txt files must exist.
}
\examples{
compute.selection.by.region(output.folder='./examples/', output.prefix='TCGA.CHOL')
}
