% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_functions_by_type.R
\name{compute_xstar}
\alias{compute_xstar}
\title{compute_xstar}
\usage{
compute_xstar(S, S2, r, s, t, Qr, Qs, Qt)
}
\arguments{
\item{S}{Numeric vector. Cumulative sum of data (e.g., \code{cumsum(y)} with a prepended 0).}

\item{S2}{Numeric vector. Cumulative sum of squares (e.g., \code{cumsum(y^2)} with a prepended 0).}

\item{r}{First index}

\item{s}{Second index}

\item{t}{Final index}

\item{Qr}{The value of the global cost at r}

\item{Qs}{The value of the global cost at s}

\item{Qt}{The value of the global cost at t}
}
\value{
The optimal value \eqn{x^\star} as a numeric scalar. Returns 0 if \code{Qr == Inf}.
}
\description{
Computes the critical point \eqn{x^\star}
This function uses empirical variances and mean differences between segments \eqn{[r,s]} and \eqn{[s,t]}
}
