% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-functions.dust.R
\name{condense.type}
\alias{condense.type}
\title{Combine reciprocal mutation types}
\usage{
condense.type(obs_exp)
}
\arguments{
\item{obs_exp:}{A data frame with observed and expected counts of variants}
}
\value{
A data frame with observed and expected counts of variants after combining reciprocal mutation types
}
\description{
Combine reciprocal mutation types
}
