% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation_cost.R
\name{Cost_1D}
\alias{Cost_1D}
\title{Cost_1D}
\usage{
Cost_1D(S, a, b, model)
}
\arguments{
\item{S}{the cumulative sums (of the data, or of the squared data for \code{"variance"}) with \code{S[1] = 0}}

\item{a}{index of the beginning of the segment in \code{S}}

\item{b}{index of the end of the segment in \code{S}}

\item{model}{the model: \code{"gauss"}, \code{"poisson"}, \code{"exp"}, \code{"geom"}, \code{"bern"}, \code{"binom"}, \code{"negbin"}, \code{"variance"}}
}
\value{
the cost of the segment
}
\description{
Cost of the segment (a, b] from the cumulative sums of the data
}
\examples{
data <- dataGenerator_1D(chpts = c(300, 600), parameters = c(0.6, 0.2),
                         nbSuccess = 10, type = "negbin")
S <- c(0, cumsum(data))
Cost_1D(S, a = 1, b = 301, model = "negbin")  # data[1:300]

}
