% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cs3}
\alias{cs3}
\title{Cumulative Sum (cs3): No Copy of Input Vector to NumericVector}
\usage{
cs3(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A double value (currently always returns 0, but the cumulative sum is computed internally).
}
\description{
Computes the cumulative sum of the input vector \code{x} without copying the data into a new container.
}
\examples{
\dontrun{
  x <- c(1, 2, 3, 4, 5)
  cs3(x)  # Returns 0, but the cumulative sum is computed internally
}
}
