% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{data_normalization_MD}
\alias{data_normalization_MD}
\title{Data Normalization Function}
\usage{
data_normalization_MD(y, type = "gauss")
}
\arguments{
\item{y}{A numeric matrix representing the time series to be normalized and then segmented.}

\item{type}{A string specifying the model type for normalization.
The available options are "gauss", "exp", "poisson", "geom", "bern", "binom", "negbin", "variance".
The default is "gauss".}
}
\value{
A numeric matrix that is the normalized version of the input time series `y`, normalized row by row.
}
\description{
Normalizes the input time series data `y` according to the specified `type`.
The normalization process depends on the statistical model type, which can be one of the following:
"gauss" (Gaussian/normal distribution), "exp" (exponential distribution),
"poisson" (Poisson distribution), "geom" (geometric distribution),
"bern" (Bernoulli distribution), "binom" (binomial distribution),
"negbin" (negative binomial distribution), or "variance"
}
\examples{
# Normalize a random time series using the Gaussian model
normalized_y <- data_normalization_MD(matrix(rnorm(100), nrow = 2), type = "gauss")
}
