% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataGenerator_1D.R
\name{dataGenerator_1D}
\alias{dataGenerator_1D}
\title{dataGenerator_1D}
\usage{
dataGenerator_1D(
  chpts = 100,
  parameters = 0.5,
  sdNoise = 1,
  gamma = 1,
  nbTrials = 10,
  nbSuccess = 10,
  type = "gauss"
)
}
\arguments{
\item{chpts}{a vector of increasing change-point indices (the last value is data length)}

\item{parameters}{Vector of segment parameters. For \code{"variance"} these
are standard deviations; for \code{"exp"} they are rates.}

\item{sdNoise}{(type \code{"gauss"}) standard deviation for the noise parameter}

\item{gamma}{(type \code{"gauss"}) vector of numbers between 0 and 1 : the coefficient of the exponential decay. By default = 1 for piecewise constant signals. If one value, it is used for all segments. Otherwise we need as many values as in \code{chpts} vector.}

\item{nbTrials}{(type \code{"binom"}) number of trials}

\item{nbSuccess}{(type \code{"negbin"}) number of successes}

\item{type}{the model: \code{"gauss"}, \code{"exp"}, \code{"poisson"}, \code{"geom"}, \code{"bern"}, \code{"binom"}, \code{"negbin"}, \code{"variance"}}
}
\value{
a univariate time series following the chosen model type and parameters
}
\description{
Generating univariate time series for multiple change-point detection based on uni-parametric models of the exponential family
}
\examples{
dataGenerator_1D(chpts = c(50,100), parameters = c(0,1), sdNoise = 0.2, type = "gauss")
dataGenerator_1D(chpts = c(50,100), parameters = c(10,20), gamma = c(0.9,0.95), type = "gauss")
dataGenerator_1D(chpts = c(50,100), parameters = c(2,7), type = "exp")
dataGenerator_1D(chpts = c(50,100), parameters = c(3,5), type = "poisson")
dataGenerator_1D(chpts = c(50,100), parameters = c(0.6,0.3), type = "geom")
dataGenerator_1D(chpts = c(50,100), parameters = c(0.7,0.2), type = "bern")
dataGenerator_1D(chpts = c(50,100), parameters = c(0.7, 0.3), nbTrials = 5, type = "binom")
dataGenerator_1D(chpts = c(50,100), parameters = c(0.4,0.7), nbSuccess = 10, type = "negbin")
dataGenerator_1D(chpts = c(50,70,120,200), parameters = c(0,3,-1,1), type = "gauss")
dataGenerator_1D(chpts = c(50,100,180), parameters = c(3,1,6), type = "variance")
}
