% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generator.R
\name{dataGenerator_AR1}
\alias{dataGenerator_AR1}
\title{dataGenerator_AR1}
\usage{
dataGenerator_AR1(chpts = 100, means = 0, sdNoise = 1, rho = 0.7)
}
\arguments{
\item{chpts}{a vector of increasing change-point indices (the last value is data length)}

\item{means}{vector of successive segment parameters (as many parameters as values in \code{chpts} vector)}

\item{sdNoise}{standard deviation for the noise parameter (type \code{"gauss"})}

\item{rho}{vector of numbers between 0 and 1 : the coefficient of the exponential decay (type \code{"gauss"}). By default = 1 for piecewise constant signals. If one value, it is used for all segments. Otherwise we need as many values as in \code{chpts} vector.}
}
\value{
a time series following the AR(1) model type
}
\description{
Generating time series for multiple change-point detection with AR1 model
}
\examples{
dataGenerator_AR1(chpts = c(50,100), means = c(0,1), sdNoise = 0.2, rho = c(0.9,0.7))
}
