% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generator.R
\name{dataGenerator_MV}
\alias{dataGenerator_MV}
\title{dataGenerator_MV}
\usage{
dataGenerator_MV(chpts = 100, means = 0, sds = 1)
}
\arguments{
\item{chpts}{a vector of increasing change-point indices}

\item{means}{vector of successive segment means}

\item{sds}{vector of successive segment standard deviation}
}
\value{
a vector of (univariate) simulated data with changes in mean and variance
}
\description{
Generating data for detecting changes in mean and variance with Gaussian model
}
\examples{
dataGenerator_MV(chpts = c(30,100,120), means = c(0,1,0), sds = c(1,1,2))
}
