% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generator.R
\name{dataGenerator_Reg}
\alias{dataGenerator_Reg}
\title{dataGenerator_Reg}
\usage{
dataGenerator_Reg(
  chpts = 100,
  A = 2,
  B = -1,
  meansX = 0,
  sdX = 1,
  sdNoise = 1,
  type = "simple"
)
}
\arguments{
\item{chpts}{a vector of increasing change-point indices}

\item{A}{vector of regression coefficients A in Ax+B simple regression model}

\item{B}{vector of regression coefficients B in Ax+B simple regression model}

\item{meansX}{vector of mean values for x values generated by a Gaussian model (if one value, we copy the value to match chpts length)}

\item{sdX}{vector of standard deviation values for x values generated by a Gaussian model  (if one value, we copy the value to match chpts length)}

\item{sdNoise}{standard deviation of the Gaussian noise}

\item{type}{\code{"simple"} for the simple regression, \code{"logistic"} for the logistic regression}
}
\value{
a dataframe with time series x and y for change-points in regression of type \code{y = Ax + B + noise} or\code{P(Y = 1) = 1/(1+exp(-(Ax+B+noise)))}.  \code{x} and \code{noise} are generated by a random Gaussian model.
}
\description{
Generating data for changes in simple regression model and simple logistic regression model
}
\examples{
dataGenerator_Reg(chpts = c(40,90), A = c(2,-1),  B = c(-1,2), meansX = c(1,2), type = "simple")
dataGenerator_Reg(chpts = c(50,100), A = c(-1,1),  B = c(1,-1), type = "logistic")
}
