% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dual_1D.R
\name{dual_1D}
\alias{dual_1D}
\title{dual_1D}
\usage{
dual_1D(
  mu,
  data,
  s1,
  s2,
  t,
  type = "gauss",
  OP = FALSE,
  penalty = 2 * length(data),
  pruningOpt = 3
)
}
\arguments{
\item{mu}{value for the dual (mu btw 0 and 1 is then re-scaled)}

\item{data}{a vector of data to segment generated by \code{dataGenerator_1D}}

\item{s1}{index for q_t^s1}

\item{s2}{index for q_t^s2}

\item{t}{time step t}

\item{type}{type of cost to use: \code{"gauss"}, \code{"exp"}, \code{"poisson"}, \code{"geom"}, \code{"bern"}, \code{"binom"}, \code{"negbin"}}

\item{OP}{FALSE or TRUE, using the OP algorithm (\code{dust_R_1D}) or not to get the \code{costQ} vector, otherwise all cost set to zero}

\item{penalty}{penalty value (non-negative)}

\item{pruningOpt}{pruning option for \code{dust_R_1D}. See \code{pruningOpt} parameter in function \code{dust_R_1D}.}
}
\value{
Value on the dual function
}
\description{
Computing the value of the dual at relative position mu (mu btw 0 and 1 is then rescaled)
}
\examples{
data <- dataGenerator_1D(chpts = c(15,30), parameters = c(1,10), type = "gauss")
dual_1D(mu =  1:99/100, data = data, s1 = 4, s2 = 1, t = 10, type = "gauss", OP = TRUE)
dual_1D(mu =  1:99/100, data = data, s1 = 4, s2 = 1, t = 10, type = "gauss")
}
