% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{DUST.1D.HW}
\alias{DUST.1D.HW}
\title{DUST.1D.HW}
\usage{
DUST.1D.HW(data, penalty = NULL, model = "gauss", method = "DUST")
}
\arguments{
\item{data}{a numeric vector}

\item{penalty}{penalty value, \code{2 log(n)} by default}

\item{model}{one of "gauss", "poisson", "exp", "geom", "bern", "binom", "negbin", "variance"}

\item{method}{one of "DUST", "DUSTib", "PELT", "OP"}
}
\value{
a list with changepoints, lastIndexSet, backend, nb and costQ
}
\description{
Same as \code{\link{dust.1D}} with the Highway (SIMD) engine.
Use \code{dust.1D(..., backend = "highway")}.
}
\examples{
y <- dataGenerator_1D(chpts = c(500, 1000), parameters = c(0, 1), type = "gauss")
dust.1D(y, backend = "highway")$changepoints
}
\keyword{internal}
