% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DustObject_1D.R
\name{dust.object.1D.HW}
\alias{dust.object.1D.HW}
\title{dust.object.1D.HW}
\usage{
dust.object.1D.HW(model = "gauss", method = "DUST")
}
\arguments{
\item{model}{the model, see \code{\link{dust.object.1D}}}

\item{method}{the pruning method, see \code{\link{dust.object.1D}}}
}
\value{
a DUST 1D object (same methods as \code{\link{dust.object.1D}})
}
\description{
Highway version of \code{\link{dust.object.1D}} (scalar object if Highway is not available).
Use \code{dust.object.1D(..., backend = "highway")}.
}
\examples{
y <- dataGenerator_1D(chpts = c(400, 800), parameters = c(2, 0.5), type = "exp")
ob <- dust.object.1D(model = "exp", backend = "highway")
ob$dust(y, 2 * log(800))$changepoints
}
\keyword{internal}
