% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dust-package.R
\docType{package}
\name{dust}
\alias{dust}
\alias{dust-package}
\title{dust: Fast Multiple Change-Point Detection}
\description{
The \pkg{dust} package implements DUST pruning for multiple change-point
detection. Its \code{dust.1D} and \code{dust.MD} interfaces support eight
one-parameter cost models for univariate and independent multivariate data.
The \code{dust.meanVar} interface detects changes in both the mean and
variance of a Gaussian series. All three offer scalar and optional Highway
backends and online objects for incremental analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/vrunge/dust}
  \item Report bugs at \url{https://github.com/vrunge/dust/issues}
}

}
\author{
\strong{Maintainer}: Vincent Runge \email{vincent.runge@univ-evry.fr}

Authors:
\itemize{
  \item Vincent Runge \email{vincent.runge@univ-evry.fr}
  \item Simon Querné \email{simon.querne@gmail.com}
  \item Félix Laplante \email{felixlaplante0@proton.me}
}

Other contributors:
\itemize{
  \item Charles Truong \email{ctruong@ens-paris-saclay.fr} [contributor]
}

}
