% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{DUST_1D}
\alias{DUST_1D}
\title{MyModule: Exposing DUST_1D to R}
\description{
This module exposes the \code{DUST_1D} C++ class to R, allowing you to create
instances of \code{DUST_1D} and call its methods directly from R.
}
