% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{DUST_2D}
\alias{DUST_2D}
\title{MyModule: Exposing DUST_2D to R}
\description{
This module exposes the \code{DUST_2D} C++ class to R, allowing you to create
instances of \code{DUST_2D} and call its methods directly from R.
}
