% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{dust_data}
\alias{dust_data}
\title{Process data for dust}
\usage{
dust_data(object, name_step = "step")
}
\arguments{
\item{object}{An object, at this point must be a \link{data.frame}}

\item{name_step}{The name of the data column within \code{object}; this
column must be integer-like and every integer must be
nonnegative and unique}
}
\description{
Prepare data for use with the \verb{$set_data()} method. This is not
required for use but tries to simplify the most common use case
where you have a \link{data.frame} with some column indicating "model
step" (\code{name_step}), and other columns that might be use in your
\code{data_compare} function. Each row will be turned into a named R
list, which your \code{dust_data} function can then work with to get
this time-steps values.
}
\details{
The data object as accepted by \code{data_set} must be a \link{list} and
each element must itself be a list with two elements; the model
\code{step} at which the data applies and any R object that corresponds
to data at that point. We expect that most of the time this second
element will be a key-value list with scalar keys, but more
flexibility may be required.
}
\examples{
d <- data.frame(step = seq(0, 50, by = 10), a = runif(6), b = runif(6))
dust::dust_data(d)
}
