% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{dust_example}
\alias{dust_example}
\title{Access dust's built-in examples.}
\usage{
dust_example(name)
}
\arguments{
\item{name}{The name of the example to use. There are four
examples: \code{sir}, \code{variable}, \code{volatility} and \code{walk} (see Details).}
}
\value{
A \code{\link{dust}} object that can be used to create a model.  See
examples for usage.
}
\description{
Access dust's built-in examples. These are compiled into the
package so that examples and tests can be run more quickly without
having to compile code directly via \code{\link[=dust]{dust()}}.  These examples are
all "toy" examples, being small and fast to run.
}
\details{
\itemize{
\item \code{sir}: a basic SIR (Susceptible, Infected, Resistant)
epidemiological model.  Draws from the binomial distribution to
update the population between each step.
\item \code{volatility}: A volatility model that might be applied to
currency fluctuations etc.
\item \code{walk}: A 1D random walk, following a Gaussian distribution each
time step.
}
}
\examples{

# A SIR (Susceptible, Infected, Resistant) epidemiological model
sir <- dust::dust_example("sir")
sir

# Initialise the model at step 0 with 50 independent trajectories
mod <- sir$new(list(), 0, 50)

# Run the model for 400 steps, collecting "infected" every 4th step
steps <- seq(0, 400, by = 4)
mod$set_index(2L)
y <- mod$simulate(steps)

# A plot of our epidemic
matplot(steps, t(drop(y)), type = "l", lty = 1, col = "#00000044",
        las = 1, xlab = "Step", ylab = "Number infected")
}
