% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{DUST_MD}
\alias{DUST_MD}
\title{MyModule: Exposing DUST_MD to R}
\description{
This module exposes the \code{DUST_MD} C++ class to R, allowing you to create
instances of \code{DUST_MD} and call its methods directly from R.
}
