% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{DUST_meanVar}
\alias{DUST_meanVar}
\title{MyModule: Exposing DUST_meanVar to R}
\description{
This module exposes the \code{DUST_meanVar} C++ class to R, allowing you to create
instances of \code{DUST_meanVar} and call its methods directly from R.
}
