% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DUST_module_classes.R
\docType{class}
\name{DUST_module_classes}
\alias{DUST_module_classes}
\title{Internal S4 classes for DUST C++ modules}
\description{
Classes generated by the Rcpp modules (1D data). Use \code{\link{dust.1D}} and \code{\link{dust.object.1D}}.
}
\seealso{
\code{\link{Rcpp_DUST_1D-class}}, \code{\link{C++Object-class}}
}
\keyword{internal}
