% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode_control.R
\name{dust_ode_control}
\alias{dust_ode_control}
\title{Create a dust_ode_control object.}
\usage{
dust_ode_control(
  max_steps = NULL,
  atol = NULL,
  rtol = NULL,
  step_size_min = NULL,
  step_size_max = NULL,
  debug_record_step_times = NULL
)
}
\arguments{
\item{max_steps}{Maxmimum number of steps to take. If the
integration attempts to take more steps that this, it will
throw an error, stopping the integration.}

\item{atol}{The per-step absolute tolerance.}

\item{rtol}{The per-step relative tolerance.  The total accuracy
will be less than this.}

\item{step_size_min}{The minimum step size.  The actual minimum
used will be the largest of the absolute value of this
\code{step_size_min} or \code{.Machine$double.eps}. If the
integration attempts to make a step smaller than this, it will
throw an error, stopping the integration.}

\item{step_size_max}{The largest step size.  By default there is
no maximum step size (Inf) so the solver can take as large a
step as it wants to.  If you have short-lived fluctuations in
your rhs that the solver may skip over by accident, then specify
a smaller maximum step size here.}

\item{debug_record_step_times}{Logical, indicating if we should record
the steps taken. This information will be available as part of
the \code{statistics()} output}
}
\value{
A named list of class "dust_ode_control"
}
\description{
Create a control object for controlling the adaptive stepper for
systems of ordinary differential equations (ODEs). The returned
object can be passed into a continuous-time dust model on
initialisation.
}
