% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openmp.R
\name{dust_openmp_support}
\alias{dust_openmp_support}
\title{Information about OpenMP support}
\usage{
dust_openmp_support(check_compile = FALSE)
}
\arguments{
\item{check_compile}{Logical, indicating if we should check if we
can compile an openmp program - this is slow the first time.}
}
\description{
Return information about OpenMP support for this system.  For
individual models look at the \verb{$has_openmp()} method
}
\examples{
# System wide support
dust::dust_openmp_support()

# Support compiled into a generator
walk <- dust::dust_example("walk")
walk$public_methods$has_openmp()

# Support from an instance of that model
model <- walk$new(list(sd = 1), 0, 1)
model$has_openmp()
}
\seealso{
\code{\link[=dust_openmp_threads]{dust_openmp_threads()}} for setting a polite number of
threads.
}
