% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{dust_package}
\alias{dust_package}
\title{Create dust model in package}
\usage{
dust_package(path, quiet = FALSE, gpu = FALSE)
}
\arguments{
\item{path}{Path to the package}

\item{quiet}{Passed to \code{cpp11::cpp_register}, if \code{TRUE} suppresses
informational notices about updates to the cpp11 files}

\item{gpu}{Create GPU code (currently must be shared across all models)}
}
\value{
Nothing, this function is called for its side effects
}
\description{
Updates a dust model in a package. The user-provided code is
assumed to the in \code{inst/dust} as a series of C++ files; a file
\code{inst/dust/model.cpp} will be transformed into a file
\code{src/model.cpp}.
}
\details{
If your code provides a class \code{model} then dust will create C++
functions such as \code{dust_model_alloc} - if your code also includes
names such as this, compilation will fail due to duplicate
symbols.

We add "cpp11 attributes" to the created functions, and will run
\code{\link[cpp11:cpp_register]{cpp11::cpp_register()}} on them once the generated code
has been created.

Your package needs a \code{src/Makevars} file to enable openmp (if your
system supports it). If it is not present then a suitable Makevars
will be written, containing\preformatted{PKG_CXXFLAGS=$(SHLIB_OPENMP_CXXFLAGS)
PKG_LIBS=$(SHLIB_OPENMP_CXXFLAGS)
}

following "Writing R Extensions" (see section "OpenMP support").
If your package does contain a \code{src/Makevars} file we do not
attempt to edit it but will error if it looks like it does not
contain these lines or similar.
}
