% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DUST_partitioning.R
\name{DUST_partitioning}
\alias{DUST_partitioning}
\title{DUST_partitioning}
\usage{
DUST_partitioning(
  values,
  penalty = 2 * log(length(values)),
  cost_function = gauss_cost,
  cost_scale = 1,
  pruning = 2,
  test_function = simple_test
)
}
\arguments{
\item{values}{a vector containing the data points}

\item{penalty}{the value of the changepoint cost penalty}

\item{cost_function}{the cost function to be used when computing the cost of the model. defaults to gauss_cost}

\item{cost_scale}{value of the factor applied to the cost function. defaults to 1}

\item{pruning}{a value in 0 = no pruning step, 1 = PELT pruning step, 2 = DUST pruning step, 3 = PELT -> DUST pruning steps}

\item{test_function}{the test function to be used when performing the DUST pruning step. defaults to simple_test}
}
\value{
a list containing a vector of the optimal cost at each step and a vector containing the optimal model on the full data
}
\description{
Computes the exact optimal partitioning for given data, cost function and cost penalization with pruning steps option (DUST as default)
}
