% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dust_2param.R
\name{dust_R_2param}
\alias{dust_R_2param}
\title{dust_R_2param}
\usage{
dust_R_2param(data, penalty, type = "meanVarExact", pruningOpt = 2)
}
\arguments{
\item{data}{a vector of data (type \code{"meanVar"}) generated by \code{dataGenerator_meanVar} or a data frame of data of dimension n x 2 (type \code{"regression"}) generated by \code{dataGenerator_Reg} with column named \code{x} and \code{y}.}

\item{penalty}{penalty value (non-negative)}

\item{type}{type of cost to use: \code{"meanVar"}, \code{"regression"}}

\item{pruningOpt}{the pruning option to use. 0 = nothing, 1 = PELT, 2 = dust, 3 = PELT + dust}
}
\value{
a list with (1) the change-point elements (each last index of each segment in \code{changepoints}), (2) a vector `\code{nb} saving the number of non-pruned elements at each iteration, (3) a vector \code{lastIndexSet} containing the non-pruned indices at the end of the algo and (4) a vector \code{costQ} saving the optimal cost at each time step.
}
\description{
DUST algorithm for univariate time-series (with different possible data models)
}
\examples{
data <- dataGenerator_meanVar(c(50,100), c(0,1), c(0.5,0.2))
dust_R_2param(data, 4*log(100), type = "meanVar", pruningOpt = 1)
dust_R_2param(data, 4*log(100), type = "meanVar", pruningOpt = 2)

data <- dataGenerator_Reg(chpts = c(50,100), A = c(1,-1), B = c(2,0))
dust_R_2param(data, 4*log(100), type = "regression", pruningOpt = 1)
dust_R_2param(data, 4*log(100), type = "regression", pruningOpt = 2)
}
