% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dust_2param.R
\name{dust_R_2param_meanVarExact}
\alias{dust_R_2param_meanVarExact}
\title{dust_R_2param_meanVarExact}
\usage{
dust_R_2param_meanVarExact(data, penalty)
}
\arguments{
\item{data}{a vector of data (type \code{"meanVar"}) generated by \code{dataGenerator_meanVar}}

\item{penalty}{penalty value (non-negative)}
}
\description{
DUST algorithm for meanVar problem
}
