% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dust_2param.R
\name{dust_R_2param_regression}
\alias{dust_R_2param_regression}
\title{dust_R_2param_regression}
\usage{
dust_R_2param_regression(data, penalty, pruningOpt)
}
\arguments{
\item{data}{a data frame of data of dimension n x 2 (type \code{"regression"}) generated by \code{dataGenerator_Reg} with column named \code{x} and \code{y}.}

\item{penalty}{penalty value (non-negative)}

\item{pruningOpt}{the pruning option to use. 0 = nothing, 1 = PELT, 2 = dust, 3 = PELT + dust}
}
\description{
DUST algorithm for regression problem
}
