% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_dual_MulltiD.R
\name{dust_R_MD}
\alias{dust_R_MD}
\title{dust_R_MD}
\usage{
dust_R_MD(
  data,
  penalty = 2 * nrow(data) * log(ncol(data)),
  type = "gauss",
  pruningOpt = 3
)
}
\arguments{
\item{data}{a matrix of data to segment generated by \code{dataGenerator_MD}}

\item{penalty}{penalty value (non-negative)}

\item{type}{type of cost to use: \code{"gauss"}, \code{"exp"}, \code{"poisson"}, \code{"geom"}, \code{"bern"}, \code{"binom"}, \code{"negbin"}}

\item{pruningOpt}{the pruning option to use. 0 = nothing, 1 = PELT, 2 = dust, 3 = PELT + dust}
}
\value{
a list with (1) the change-point elements (each last index of each segment in \code{changepoints}), (2) a vector `\code{nb} saving the number of non-pruned elements at each iteration, (3) a vector \code{lastIndexSet} containing the non-pruned indices at the end of the algo and (4) a vector \code{costQ} saving the optimal cost at each time step.
}
\description{
DUST algorithm for multivariate time-series (with different possible data models)
}
\examples{
data <- dataGenerator_MD(chpts = c(100,200),
                         parameters = data.frame(ts1 = c(0,-2), ts2 = c(0,1)))
dust_R_MD(data)$changepoints
}
