% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{DUST_Reg}
\alias{DUST_Reg}
\title{MyModule: Exposing DUST_Reg to R}
\description{
This module exposes the \code{DUST_Reg} C++ class to R, allowing you to create
instances of \code{DUST_Reg} and call its methods directly from R.
}
