% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{dust_repair_environment}
\alias{dust_repair_environment}
\title{Repair dust environment}
\usage{
dust_repair_environment(generator, quiet = FALSE)
}
\arguments{
\item{generator}{The dust generateor}

\item{quiet}{Logical, indicating if we should be quiet (default
prints some progress information)}
}
\value{
Nothing, called for its side effects
}
\description{
Repair the environment of a dust object created by [\link{dust}]
and then saved and reloaded by [\link{saveRDS}] and
[\link{readRDS}]. Because we use a fake temporary package to hold the
generated code, it will not ordinarily be loaded properly without
using this.
}
