% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rng.R
\name{dust_rng}
\alias{dust_rng}
\title{Dust Random Number Generator}
\value{
A \code{dust_rng} object, which can be used to drawn random
numbers from dust's distributions.
}
\description{
Create an object that can be used to generate random
numbers with the same RNG as dust uses internally.  This is
primarily meant for debugging and testing.
}
\examples{
rng <- dust::dust_rng$new(42)

# Shorthand for Uniform(0, 1)
rng$unif_rand(5)

# Uniform random numbers between min and max
rng$uniform(5, -2, 6)

# Normally distributed random numbers with mean and sd
rng$normal(5, 4, 2)

# Binomially distributed random numbers with size and prob
rng$binomial(5, 10, 0.3)

# Poisson distributed random numbers with mean lambda
rng$poisson(5, 2)

# Exponentially distributed random numbers with rate
rng$exponential(5, 2)
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{info}}{Information about the generator (read-only)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{dust_rng$new()}}
\item \href{#method-size}{\code{dust_rng$size()}}
\item \href{#method-real_type}{\code{dust_rng$real_type()}}
\item \href{#method-jump}{\code{dust_rng$jump()}}
\item \href{#method-long_jump}{\code{dust_rng$long_jump()}}
\item \href{#method-unif_rand}{\code{dust_rng$unif_rand()}}
\item \href{#method-uniform}{\code{dust_rng$uniform()}}
\item \href{#method-normal}{\code{dust_rng$normal()}}
\item \href{#method-binomial}{\code{dust_rng$binomial()}}
\item \href{#method-poisson}{\code{dust_rng$poisson()}}
\item \href{#method-exponential}{\code{dust_rng$exponential()}}
\item \href{#method-state}{\code{dust_rng$state()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a \code{dust_rng} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dust_rng$new(
  seed,
  n_generators = 1L,
  real_type = "double",
  deterministic = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{seed}}{The seed, as an integer or as a raw vector.}

\item{\code{n_generators}}{The number of generators to use. While this
function never runs in parallel, this is used to create a set of
interleaved independent generators as dust would use in a model.}

\item{\code{real_type}}{The type of floating point number to use. Currently
only \code{float} and \code{double} are supported (with \code{double} being
the default). This will have no (or negligible) impact on speed,
but exists to test the low-precision generators.}

\item{\code{deterministic}}{Logical, indicating if we should use
"deterministic" mode where distributions return their
expectations and the state is never changed.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-size"></a>}}
\if{latex}{\out{\hypertarget{method-size}{}}}
\subsection{Method \code{size()}}{
Number of generators available
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dust_rng$size()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-real_type"></a>}}
\if{latex}{\out{\hypertarget{method-real_type}{}}}
\subsection{Method \code{real_type()}}{
Indicates the floating point type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dust_rng$real_type()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jump"></a>}}
\if{latex}{\out{\hypertarget{method-jump}{}}}
\subsection{Method \code{jump()}}{
The jump function for the generator, equivalent to
2^128 numbers drawn from the generator.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dust_rng$jump()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-long_jump"></a>}}
\if{latex}{\out{\hypertarget{method-long_jump}{}}}
\subsection{Method \code{long_jump()}}{
The \code{long_jump} function for the generator, equivalent
to 2^192 numbers drawn from the generator.
Generate \code{n} numbers from a standard uniform distribution
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dust_rng$long_jump()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-unif_rand"></a>}}
\if{latex}{\out{\hypertarget{method-unif_rand}{}}}
\subsection{Method \code{unif_rand()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dust_rng$unif_rand(n)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{Number of samples to draw
Generate \code{n} numbers from a uniform distribution}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-uniform"></a>}}
\if{latex}{\out{\hypertarget{method-uniform}{}}}
\subsection{Method \code{uniform()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dust_rng$uniform(n, min, max)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{Number of samples to draw}

\item{\code{min}}{The minimum of the distribution (length 1 or n)}

\item{\code{max}}{The maximum of the distribution (length 1 or n)
Generate \code{n} numbers from a normal distribution}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-normal"></a>}}
\if{latex}{\out{\hypertarget{method-normal}{}}}
\subsection{Method \code{normal()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dust_rng$normal(n, mean, sd)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{Number of samples to draw}

\item{\code{mean}}{The mean of the distribution (length 1 or n)}

\item{\code{sd}}{The standard deviation of the distribution (length 1 or n)
Generate \code{n} numbers from a binomial distribution}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-binomial"></a>}}
\if{latex}{\out{\hypertarget{method-binomial}{}}}
\subsection{Method \code{binomial()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dust_rng$binomial(n, size, prob)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{Number of samples to draw}

\item{\code{size}}{The number of trials (zero or more, length 1 or n)}

\item{\code{prob}}{The probability of success on each trial
(between 0 and 1, length 1 or n)
Generate \code{n} numbers from a Poisson distribution}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-poisson"></a>}}
\if{latex}{\out{\hypertarget{method-poisson}{}}}
\subsection{Method \code{poisson()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dust_rng$poisson(n, lambda)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{Number of samples to draw}

\item{\code{lambda}}{The mean (zero or more, length 1 or n)
Generate \code{n} numbers from a exponential distribution}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-exponential"></a>}}
\if{latex}{\out{\hypertarget{method-exponential}{}}}
\subsection{Method \code{exponential()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dust_rng$exponential(n, rate)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{Number of samples to draw}

\item{\code{rate}}{The rate of the exponential}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-state"></a>}}
\if{latex}{\out{\hypertarget{method-state}{}}}
\subsection{Method \code{state()}}{
Returns the state of the random number generator. This returns a
raw vector of length 32 * n_generators. It is primarily intended for
debugging as one cannot (yet) initialise a dust_rng object with this
state.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dust_rng$state()}\if{html}{\out{</div>}}
}

}
}
