% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rng_distributed.R
\name{dust_rng_distributed_state}
\alias{dust_rng_distributed_state}
\alias{dust_rng_distributed_pointer}
\title{Create a set of distributed seeds}
\usage{
dust_rng_distributed_state(
  seed = NULL,
  n_streams = 1L,
  n_nodes = 1L,
  algorithm = "xoshiro256plus"
)

dust_rng_distributed_pointer(
  seed = NULL,
  n_streams = 1L,
  n_nodes = 1L,
  algorithm = "xoshiro256plus"
)
}
\arguments{
\item{seed}{Initial seed to use. As for \link{dust_rng}, this can
be \code{NULL} (create a seed using R's generators), an integer or a
raw vector of appropriate length.}

\item{n_streams}{The number of streams to create per node. If
passing the results of this seed to a dust object's initialiser
(see \link{dust_generator}) you can safely leave this at 1, but
if using in a standalone setting, and especially if using
\code{dust_rng_distributed_pointer}, you may need to set this to the
appropriate length.}

\item{n_nodes}{The number of separate seeds to create. Each will
be separated by a "long jump" for your generator.}

\item{algorithm}{The name of an algorithm to use. Alternatively
pass a \code{dust_generator} or \code{dust} object here to select the
algorithm used by that object automatically.}
}
\value{
A list of either raw vectors (for
\code{dust_rng_distributed_state}) or of \link{dust_rng_pointer}
objects (for \code{dust_rng_distributed_pointer})
}
\description{
Create a set of initial random number seeds suitable for using
within a distributed context (over multiple processes or nodes) at
a level higher than a single group of synchronised threads.
}
\details{
See \code{vignette("rng_distributed")} for a proper introduction to
these functions.
}
\examples{
dust::dust_rng_distributed_state(n_nodes = 2)
dust::dust_rng_distributed_pointer(n_nodes = 2)
}
