% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rng_pointer.R
\name{dust_rng_pointer}
\alias{dust_rng_pointer}
\title{Create pointer to random number generator

For external use, vignette coming soon.}
\description{
Create pointer to random number generator

For external use, vignette coming soon.

Create pointer to random number generator

For external use, vignette coming soon.
}
\examples{
dust::dust_rng_pointer$new()
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{algorithm}}{The name of the generator algorithm used (read-only)}

\item{\code{n_streams}}{The number of streams of random numbers provided
(read-only)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{dust_rng_pointer$new()}}
\item \href{#method-sync}{\code{dust_rng_pointer$sync()}}
\item \href{#method-state}{\code{dust_rng_pointer$state()}}
\item \href{#method-is_current}{\code{dust_rng_pointer$is_current()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{dust_rng_pointer} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dust_rng_pointer$new(seed = NULL, n_streams = 1L, algorithm = "xoshiro256plus")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{seed}}{The random number seed to use (see \link{dust_rng}
for details)}

\item{\code{n_streams}}{The number of independent random number streams to
create}

\item{\code{algorithm}}{The random number algorithm to use. The default is
\code{xoshiro256plus} which is a good general choice}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sync"></a>}}
\if{latex}{\out{\hypertarget{method-sync}{}}}
\subsection{Method \code{sync()}}{
Synchronise the R copy of the random number state.
Typically this is only needed before serialisation if you have
ever used the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dust_rng_pointer$sync()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-state"></a>}}
\if{latex}{\out{\hypertarget{method-state}{}}}
\subsection{Method \code{state()}}{
Return a raw vector of state. This can be used to
create other generators with the same state.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dust_rng_pointer$state()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-is_current"></a>}}
\if{latex}{\out{\hypertarget{method-is_current}{}}}
\subsection{Method \code{is_current()}}{
Return a logical, indicating if the random number
state that would be returned by \code{state()} is "current" (i.e., the
same as the copy held in the pointer) or not. This is \code{TRUE} on
creation or immediately after calling \verb{$sync()} and \code{FALSE} after
any use of the pointer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dust_rng_pointer$is_current()}\if{html}{\out{</div>}}
}

}
}
