% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_functions_by_type.R
\name{evalDual}
\alias{evalDual}
\title{evalDual}
\usage{
evalDual(mu, A, B, S, s1, s2, t, const1, const2)
}
\arguments{
\item{mu}{the mu value, evaluation point}

\item{A}{the nonlinear function}

\item{B}{the B derived from A function}

\item{S}{The cumsum S vector}

\item{s1}{the first index}

\item{s2}{the second index}

\item{t}{the time step}

\item{const1}{first constant term}

\item{const2}{second constant term}
}
\description{
evaluation of the dual function
}
