% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_functions_by_type.R
\name{evalDual_meanVar2}
\alias{evalDual_meanVar2}
\title{evalDual_meanVar2}
\usage{
evalDual_meanVar2(x, S, S2, r, s, t, Qr, Qs, Qt)
}
\arguments{
\item{x}{The evaluation point}

\item{S}{Cumulative sum vector of data}

\item{S2}{Cumulative sum vector of squared data}

\item{r}{First index}

\item{s}{Second index}

\item{t}{Final index}

\item{Qr}{The value of the global cost at r}

\item{Qs}{The value of the global cost at s}

\item{Qt}{The value of the global cost at t}
}
\description{
Evaluate the D(x) function as defined.
}
