% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_functions_by_type.R
\name{evalDual_regression}
\alias{evalDual_regression}
\title{evalDual_regression}
\usage{
evalDual_regression(mu, A, B, C, D, E, f, s1, s2, t, const1, const2)
}
\arguments{
\item{mu}{mu}

\item{A}{A}

\item{B}{B}

\item{C}{C}

\item{D}{D}

\item{E}{E}

\item{f}{f}

\item{s1}{the first index}

\item{s2}{the second index}

\item{t}{the time step}

\item{const1}{first constant term}

\item{const2}{second constant term}
}
\description{
evaluation of the dual function
}
