% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_functions_by_type.R
\name{evalPrimal}
\alias{evalPrimal}
\title{evalPrimal}
\usage{
evalPrimal(nat_theta, A, data, const)
}
\arguments{
\item{nat_theta}{the value of the natural parameter theta}

\item{A}{the nonlinear function}

\item{data}{the data to use in sum}

\item{const}{the constant term}
}
\description{
Evaluation of the primal function
}
