% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-functions.dust.R
\name{find.outliers}
\alias{find.outliers}
\title{Find hypo- and hyper-mutated samples}
\usage{
find.outliers(input.file.name, output.folder, output.prefix)
}
\arguments{
\item{input.file.name:}{A VCF-formatted file to read SnpEff annotated somatic variants. This input file shall contain these fields: Tumor_Sample_Barcode, Chromosome, Start_Position, dbSNP_RS, Reference_Allele, Tumor_Seq_Allele2, FILTER, One_Consequence, Hugo_Symbol, Gene, Feature, ENSP, HGVSc, HGVSp_Short, Amino_acids, Codons, ENSP, RefSeq, Entrez_Gene_Id.}

\item{output.folder:}{the folder to write the output files}

\item{output.prefix:}{prefix used to name the output file as prefix.outlier.txt.}
}
\description{
Find hypo- and hyper-mutated samples
}
\examples{
find.outliers(input.file.name='./examples/TCGA.CHOL.somatic.maf.gz', output.folder='./examples/', output.prefix='TCGA.CHOL');
}
