% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-functions.dust.R
\name{find.sig.selection}
\alias{find.sig.selection}
\title{Filter the CDD regions for those under significant positive selection. It requires the output.folder contains files from running the parse.aggregated.mut(), annotate.by.cdd.region(), and compute.selection.by.region() functions. That is, prefix.mut.filtered.txt, prefix.mut.region.txt, and prefix.selection.region.txt files must exist.}
\usage{
find.sig.selection(
  output.folder = output.folder,
  output.prefix = output.prefix,
  N.sim = 1000,
  fp.cutoff.rate = 0.1,
  cnt.cutoff.rate = NULL,
  cnt.mis.cutoff = 6,
  cnt.non.cutoff = 6,
  cnt.both.cutoff = 6,
  sel.cutoff = 1
)
}
\arguments{
\item{output.folder:}{the folder to write output files}

\item{output.prefix:}{prefix used to name the output file as prefix.sig.region.txt.}

\item{N.sim:}{the number of simulations performed when calling the compute.selection.by.region() function. Default value is 1,000.}

\item{fp.cutoff.rate:}{false positive cutoff rate (default: 0.1).}

\item{cnt.non.cutoff:}{the minimum number of tumors harboring proteinitruncating mutations in a CDD region (default: 6).}

\item{cnt.mis.cutoff:}{the minimum number of tumors harboring missense or protein-truncating mutations in a CDD region (default: 6).}

\item{cnt.cutoff.rate:}{the minimum fraction of tumors harboring mutations in the CDD region (default: NULL). If set (range: 0 - 1), the minimum number of tumors will be calculated as round(cnt.cutoff.rate * number of samples in the prefix.mut.filtered.txt file. If this number is larger than cnt.both.cutoff, it will overwrite cnt.both.cutoff, cnt.mis.cutoff, and cnt.non.cutoff. If not set (NULL), cnt.mis.cutoff, cnt.non.cutoff, and cnt.both.cutoff must be set.}

\item{sel.cutoff:}{the minimum selection coefficient of a CDD region (default: 1).}
}
\description{
Filter the CDD regions for those under significant positive selection. It requires the output.folder contains files from running the parse.aggregated.mut(), annotate.by.cdd.region(), and compute.selection.by.region() functions. That is, prefix.mut.filtered.txt, prefix.mut.region.txt, and prefix.selection.region.txt files must exist.
}
\examples{
find.sig.selection(output.folder='./examples/', output.prefix='TCGA.CHOL')
}
