% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{flat_OP_MD}
\alias{flat_OP_MD}
\title{Optimal Partitioning in MD with Flat Model}
\usage{
flat_OP_MD(inData, inPenalty = NULL)
}
\arguments{
\item{inData}{A numeric matrix representing the input data for partitioning.}

\item{inPenalty}{An optional numeric penalty parameter to control the number of partitions. Defaults to \code{NULL}, indicating a default penalty is used.}
}
\value{
A list containing the results of the optimal partitioning, including identified change points and other model details.
}
\description{
Computes the optimal partitioning of multi-dimensional data using a flat model with an optional penalty parameter.
}
