% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costFunctions.R
\name{gauss_cost}
\alias{gauss_cost}
\title{gauss_cost}
\usage{
gauss_cost(loc, loc_current, values_cumsum)
}
\arguments{
\item{loc}{first index of the sum}

\item{loc_current}{last index of the sum}

\item{values_cumsum}{cumulative sum of the data}
}
\value{
the linear and quadratic terms of the L2 cost function with optimal parameter
}
\description{
Computes the variable part of the L2 cost with optimal parameter on given data
}
