% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generator.R
\name{generate_random_changepoints}
\alias{generate_random_changepoints}
\title{generate_random_changepoints}
\usage{
generate_random_changepoints(n, K, mu, data_variance, changepoint_variance)
}
\arguments{
\item{n}{size of the data}

\item{K}{number of changepoints to be created}

\item{mu}{either the general mean of the data or a vector of the means of each segment (K + 1)}

\item{data_variance}{noise variance}

\item{changepoint_variance}{variance of the uniform distribution deciding the size of the segment jumps}
}
\value{
a univariate time series following the input parameters
}
\description{
Generates 1D data of size n with K random changepoints
}
\examples{
generate_random_changepoints(1e3, 1e3/50, 20, 1, 9)
generate_random_changepoints(1e2, 0, 20, 1, 9)
}
