% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_functions_by_type.R
\name{min_cost}
\alias{min_cost}
\title{min_cost}
\usage{
min_cost(A, B, S, s, t, const)
}
\arguments{
\item{A}{the nonlinear function}

\item{B}{the B derived from A function}

\item{S}{the cumsum data}

\item{s}{the start index}

\item{t}{the end index}

\item{const}{the constant term}
}
\description{
minimal cost value
}
