% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_functions_by_type.R
\name{min_cost_meanVar}
\alias{min_cost_meanVar}
\title{min_cost_meanVar}
\usage{
min_cost_meanVar(S, S2, s, t, const)
}
\arguments{
\item{S}{The cumsum S vector}

\item{S2}{The second cumsum vector for squared data}

\item{s}{the start index}

\item{t}{the end step}

\item{const}{constant term}
}
\description{
evaluation of the cost function at its minimum for mean and var change Gaussian problem
}
