% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_functions_by_type.R
\name{min_cost_regression}
\alias{min_cost_regression}
\title{min_cost_regression}
\usage{
min_cost_regression(A, B, C, D, E, f, s, t, const)
}
\arguments{
\item{A}{A}

\item{B}{B}

\item{C}{C}

\item{D}{D}

\item{E}{E}

\item{f}{f}

\item{s}{the start index}

\item{t}{the end index}

\item{const}{constant term}
}
\description{
evaluation of the min cost value
}
