% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_functions_by_type.R
\name{mu_max}
\alias{mu_max}
\title{mu_max}
\usage{
mu_max(S, s1, s2, t, type)
}
\arguments{
\item{S}{The cumsum S vector}

\item{s1}{the first index}

\item{s2}{the second index}

\item{t}{the time step}

\item{type}{the type of cost}
}
\description{
Getting the argmax dual value
}
