% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_functions_by_type.R
\name{mu_max_2param}
\alias{mu_max_2param}
\title{mu_max_2param}
\usage{
mu_max_2param(S, S2, s1, s2, t)
}
\arguments{
\item{S}{The cumsum S vector}

\item{S2}{The second cumsum vector for squared data}

\item{s1}{the first index}

\item{s2}{the second index}

\item{t}{the time step}
}
\description{
evaluation of the mu max in 2 param case
}
