% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{OP.gauss.1D}
\alias{OP.gauss.1D}
\title{MyModule: Exposing OP.gauss.1D to R}
\description{
This module exposes the \code{OP.gauss.1D} C++ class to R, allowing you to create
instances of \code{OP.gauss.1D} and call its methods directly from R.
}
