% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{OP.gauss.MD}
\alias{OP.gauss.MD}
\title{MyModule: Exposing OP.gauss.MD to R}
\description{
This module exposes the \code{OP.gauss.MD} C++ class to R, allowing you to create
instances of \code{OP.gauss.MD} and call its methods directly from R.
}
