% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_OP.R
\name{OP_R_2param}
\alias{OP_R_2param}
\title{OP_R_2param}
\usage{
OP_R_2param(data, penalty, type = "meanVar")
}
\arguments{
\item{data}{a vector of data (type \code{"meanVar"}) generated by \code{dataGenerator_meanVar} or a data frame of data of dimension n x 2 (type \code{"regression"}) generated by \code{dataGenerator_Reg} with column named \code{x} and \code{y}.}

\item{penalty}{penalty value (non-negative)}

\item{type}{type of cost to use: \code{"meanVar"}, \code{"regression"}}
}
\value{
a list with the change-point elements (each last index of each segment) and a vector \code{costQ} saving the optimal cost at each time step
}
\description{
OP algorithm for bi-parametric time-series (with different possible links)
}
\examples{

data <- dataGenerator_meanVar(c(50,100), c(0,1), c(0.5,0.6))
OP_R_2param(data, 4*log(100), type = "meanVar")

data <- dataGenerator_meanVar(c(100,300), c(0,0), c(1,2))
OP_R_2param(data, 4*log(100), type = "meanVar")$changepoints

data <- dataGenerator_meanVar(c(200,400,600), c(0,1,1), c(1,1,2))
OP_R_2param(data, 4*log(300), type = "meanVar")$changepoints

data <- dataGenerator_Reg(chpts = c(50,100), A = c(1,-1), B = c(2,0))
OP_R_2param(data, 4*log(100), type = "regression")

data <- dataGenerator_Reg(chpts = c(100,200), A = c(1,1), B = c(1,-1))
OP_R_2param(data, 4*log(100), type = "regression")$changepoints
}
