% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-functions.dust.R
\name{parse.aggregated.mut.2024}
\alias{parse.aggregated.mut.2024}
\title{Parse the MAF file to aggregate variants by genes and mutation types}
\usage{
parse.aggregated.mut.2024(
  input.file.name,
  output.folder = "",
  output.prefix = "",
  fa_aa = NULL,
  fa_bp = NULL,
  recover = T
)
}
\arguments{
\item{input.file.name:}{A VCF-formatted file to read SnpEff annotated somatic variants. This input file shall contain these fields: Tumor_Sample_Barcode, Chromosome, Start_Position, dbSNP_RS, Reference_Allele, Tumor_Seq_Allele2, FILTER, One_Consequence, Hugo_Symbol, Gene, Feature, ENSP, HGVSc, HGVSp_Short, Amino_acids, Codons, ENSP, RefSeq, Entrez_Gene_Id.}

\item{output.folder:}{the folder to write output files}

\item{output.prefix:}{prefix used to name the output files. Output files include prefix.error.txt (mutations cannot be mapped), prefix.mut.filtered.txt (somatic variants with matching types, prefix.mut.cnt.txt (somatic variants aggregated by sample, gene and mutational type), prefix.mut.summary.txt (somatic variants aggregated by gene and mutational type across all samples), prefix.symbol_2_cds_id.txt (mapping gene symbols to ensembl transcript ids), and prefix.log.txt (log information).}
}
\description{
Parse the MAF file to aggregate variants by genes and mutation types
}
\examples{
parse.aggregated.mut(input.file.name='./examples/TCGA.CHOL.somatic.maf.gz', output.folder='./examples/', output.prefix='TCGA.CHOL')
}
