% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-functions.dust.R
\name{plot_regions}
\alias{plot_regions}
\title{Plot mutation distribution in CDD regions and save to a PDF file. It requires the output.folder contains files from running the parse.aggregated.mut(), annotate.by.cdd.region(), and find.sig.selection() functions. That is, prefix.mut.filtered.txt, prefix.mut.region.txt, and prefix.sig.region.txt files must exist.}
\usage{
plot_regions(
  output.folder,
  output.prefix,
  selected.regions = NULL,
  pdf.file.name = "plot.regions.sig.pdf",
  rare.mut.cutoff = 3,
  plot.nrow = 3,
  plot.ncol = 4,
  main = ""
)
}
\arguments{
\item{plot.nrow}{and plot.col: the number of rows and columns in the plot (default: 3x4 grid)}

\item{pdf.file.name:}{name of the PDF file for saving the plot (default: 'plot.regions.sig.pdf')}

\item{output.folder:}{the folder to write output files}

\item{output.prefix:}{prefix used to name the prefix.mut.filtered.txt and prefix.mut.region.txt files.}

\item{selected.regions:}{a vector of CDD IDs (db_xref) to plot (default NULL). If set to NULL, all CDD regions in the prefix.sig.region.txt file will be plotted. If it is not NULL, those overalpping with the prefix.sig.region.txt file will be plotted.}

\item{rare.mut.cutoff:}{the minimum number of missense or protein-truncating mutations (default: 3) in a gene for which to be considered as "rarely" mutated. For a rarely mutated gene, its gene name will be enclosed inside a parenthesis.}

\item{main:}{strings (default: '') appended to the title of the plot, which is the CDD ID.}
}
\description{
Plot mutation distribution in CDD regions and save to a PDF file. It requires the output.folder contains files from running the parse.aggregated.mut(), annotate.by.cdd.region(), and find.sig.selection() functions. That is, prefix.mut.filtered.txt, prefix.mut.region.txt, and prefix.sig.region.txt files must exist.
}
