% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-functions.dust.R
\name{plot_this_region}
\alias{plot_this_region}
\title{Plot mutation distribution for one CDD region.}
\usage{
plot_this_region(
  target,
  mut.filtered,
  mut.cdd,
  uu,
  sig,
  rare.mut.cutoff = 3,
  main = "",
  verbose = F
)
}
\arguments{
\item{target:}{the ID (db_xref) of the CDD region to be plotted.}

\item{mut.filtered:}{the data frame from the prefix.mut.fitlered.txt file.}

\item{mut.cdd:}{the data frame from the prefix.mut.region.txt file.}

\item{uu:}{the vector of unique CDD IDs from the seq.gapped.list.}

\item{sig:}{the data frame from the prefix.sig.region.txt file}

\item{rare.mut.cutoff:}{the minimum number of missense or protein-truncating mutations (default: 3) in a gene for which to be considered as "rarely" mutated. For a rarely mutated gene, its gene name will be enclosed inside a parenthesis.}

\item{main:}{strings (default: '') appended to the title of the plot, which is the CDD ID.}

\item{verbose:}{TRUE or FALSE (default: ALSE) controls if the name of the genes in a CDD region will be printed in the standard output.}
}
\description{
Plot mutation distribution for one CDD region.
}
