% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-functions.dust.R
\name{prob}
\alias{prob}
\title{optimization function to estimate independent probabilities of a specific mutation type}
\usage{
prob(theta, obs_exp, compare)
}
\description{
optimization function to estimate independent probabilities of a specific mutation type
}
