% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pruning_step.R
\name{pruning_step}
\alias{pruning_step}
\title{pruning_step}
\usage{
pruning_step(
  values_cumsum,
  pruning = 2,
  test_function = simple_test,
  cost_scale = 1
)
}
\arguments{
\item{values_cumsum}{a vector containing the data points}

\item{pruning}{a value in 0 = no pruning step, 1 = PELT pruning step, 2 = DUST pruning step, 3 = PELT -> DUST pruning steps}

\item{test_function}{the test function to be used when performing the DUST pruning step. defaults to simple_test}

\item{cost_scale}{value of the factor applied to the cost function. defaults to 1}
}
\value{
a vector of the indices kept after the pruning step
}
\description{
Performs the pruning step over the data with given test function
}
