% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_functions_by_type.R
\name{R}
\alias{R}
\title{R}
\usage{
R(mu, S, s1, s2, t)
}
\arguments{
\item{mu}{the mu value}

\item{S}{The cumsum S vector}

\item{s1}{the first index}

\item{s2}{the second index}

\item{t}{the time step}
}
\description{
ratio
}
