% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4.R
\docType{class}
\name{Rcpp_Detector-class}
\alias{Rcpp_Detector-class}
\title{Rcpp_Detector Class}
\description{
S4 class of the change point detection object (\code{\link{dust.object.1D}}, \code{\link{dust.object.MD}}, \code{\link{dust.object.meanVar}}).
}
\keyword{internal}
