% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4.R
\docType{class}
\name{Rcpp_DUST_1D_HW_Obj-class}
\alias{Rcpp_DUST_1D_HW_Obj-class}
\title{Rcpp_DUST_1D_HW_Obj Class}
\description{
An internal S4 class for the Highway backend of
\code{\link{dust.object.1D}}.
}
\keyword{internal}
