% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtrack_changepoints.R
\name{retrieve_changepoints}
\alias{retrieve_changepoints}
\title{retrieve_changepoints}
\usage{
retrieve_changepoints(backtrack_changepoints)
}
\arguments{
\item{backtrack_changepoints}{a vector of indices. a path must lead from the last value to 0}
}
\value{
a vector of all optimal changepoints at time n = length(input)
}
\description{
builds the optimal changepoints by backtracking the optimal last changepoint path
}
\examples{
example_backtrack = c(0, 0, 0, 2, 3, 4, 0, 2)
retrieve_changepoints(example_backtrack)

example_backtrack2 = c(0, 0, 0, 2, 2, 2, 0, 0)
retrieve_changepoints(example_backtrack2)

}
