% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sdDiff}
\alias{sdDiff}
\title{sdDiff}
\usage{
sdDiff(y, method = "HALL")
}
\arguments{
\item{y}{a numeric vector}

\item{method}{\code{"HALL"} (default), \code{"MAD"} or \code{"SD"}
\itemize{
  \item \code{"HALL"}: Hall estimator (weighted differences)
  \item \code{"MAD"}: MAD of the differences
  \item \code{"SD"}: standard deviation of the differences
}}
}
\value{
the estimated standard deviation
}
\description{
Estimation of the noise standard deviation from the differences of the data (robust to the changes in mean)
}
\examples{
y <- dataGenerator_1D(chpts = c(300, 600), parameters = c(0, 1), sdNoise = 2, type = "gauss")
sdDiff(y)
sdDiff(y, "MAD")
sdDiff(y, "SD")

}
