% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DUST_testFunctions.R
\name{simple_test}
\alias{simple_test}
\title{simple_test}
\usage{
simple_test(values_cumsum, cost_scale = 1, pruning = 2)
}
\arguments{
\item{values_cumsum}{the cumulative sum of the data}

\item{cost_scale}{value of the factor applied to the cost function. defaults to 1}

\item{pruning}{a value in 0 = no pruning step, 1 = PELT pruning step, 2 = DUST pruning step, 3 = PELT -> DUST pruning steps}
}
\value{
the Boolean value of the test
}
\description{
Performs the simple lagrangian dual test in the gaussian penalized changepoint detection model
}
