% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxFunctions.R
\name{slice_cumsum}
\alias{slice_cumsum}
\title{generate_random_changepoints}
\usage{
slice_cumsum(csum, loc, loc_current)
}
\arguments{
\item{csum}{a vector of the cumulative sum of the data}

\item{loc}{lower (excluded) bound of the range over which the sum is performed}

\item{loc_current}{upper (included) bound of the range over which the sum is performed}
}
\value{
the sum of the data between the (loc+1)-th element and the (loc_current)-th element
}
\description{
Retrieves the sum between two values by using the data's cumulative sum
}
\examples{
generate_random_changepoints(1e3, 1e3/50, 20, 1, 9)
generate_random_changepoints(1e2, 0, 20, 1, 9)
}
