% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_aux_functions.R
\name{backtracking_changepoint}
\alias{backtracking_changepoint}
\title{backtracking_changepoint}
\usage{
backtracking_changepoint(cp, n)
}
\arguments{
\item{cp}{vector of changepoints of size n+1}

\item{n}{data length}
}
\description{
Shifting integer by +1 value
}
