% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{data_normalization_1D}
\alias{data_normalization_1D}
\title{data_normalization_1D}
\usage{
data_normalization_1D(y, type = "gauss", size = NA_real_)
}
\arguments{
\item{y}{a numeric vector}

\item{type}{the model: \code{"gauss"} (default), \code{"poisson"}, \code{"exp"}, \code{"geom"}, \code{"bern"}, \code{"binom"}, \code{"negbin"}, \code{"variance"}}

\item{size}{number of trials (binom) or number of successes (negbin). Required for these two models.}
}
\value{
the normalized data
}
\description{
Normalization of the data before using dust.1D with the default penalty
\itemize{
  \item \code{"gauss"}: division by \code{sdDiff(y)}
  \item \code{"poisson"}, \code{"exp"}: division by the mean
  \item \code{"binom"}, \code{"negbin"}: division by \code{size}
  \item \code{"variance"}: the mean is removed
  \item \code{"geom"}, \code{"bern"}: no change
}
}
\examples{
y <- dataGenerator_1D(chpts = c(300, 600), parameters = c(0, 1), sdNoise = 2, type = "gauss")
sdDiff(data_normalization_1D(y))

y <- dataGenerator_1D(chpts = c(300, 600), parameters = c(2, 8), type = "poisson")
mean(data_normalization_1D(y, type = "poisson"))

y <- dataGenerator_1D(chpts = c(300, 600), parameters = c(0.4, 0.7), nbTrials = 5, type = "binom")
data_normalization_1D(y, type = "binom", size = 5)[1:10]

}
