% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{DUST.1D.HW.backend}
\alias{DUST.1D.HW.backend}
\title{DUST.1D.HW.backend}
\usage{
DUST.1D.HW.backend()
}
\value{
"highway" or "scalar"
}
\description{
Backend available in this installation: "highway" or "scalar".
}
\keyword{internal}
