% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dustmeanVar.R
\name{dust.meanVar}
\alias{dust.meanVar}
\title{Run the meanVar Dust Change Point Detection}
\usage{
dust.meanVar(
  data = data,
  penalty = 2 * nrow(data) * log(length(data)),
  nbLoops = 10
)
}
\arguments{
\item{data}{A numeric matrix. The time series data on which change point detection is performed. Each row represents a time-series.}

\item{penalty}{A numeric value. The penalty applied for adding a new change point. By default, it is set to \code{4 x log(length(data))}.}

\item{nbLoops}{An integer. The number of loops to run in the max dual optimization algorithm. Default is 10.}
}
\description{
This function performs change point detection for mean and variance in univariate time series using the DUST algorithm.
}
