% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dustmeanVar.R
\name{dust.object.meanVar}
\alias{dust.object.meanVar}
\title{dust.object.meanVar}
\usage{
dust.object.meanVar(method = "1D", threads = .default_threads(method))
}
\arguments{
\item{method}{\code{"1D"} (default, one constraint), \code{"2D"} (two constraints) or \code{"PELT"}}

\item{threads}{number of threads for the scan of the indices. By default, all the cores for \code{"PELT"} (many indices), 1 otherwise.}
}
\value{
A DUST object with the methods \code{append_data(data, penalty)}, \code{update_partition()}, \code{get_partition()}, \code{get_info()} and \code{dust(data, penalty)}
}
\description{
Constructs a DUST object for changes in mean and variance, with methods to add data and update the segmentation
}
\details{
The penalty is fixed at the first call of \code{append_data} (with \code{NULL}, \code{3 log(n)} with n the size of the first data).
}
\examples{
y <- c(rnorm(80), rnorm(80, mean = 2, sd = 1.5))
ob <- dust.object.meanVar(method = "2D")
ob$append_data(y[1:80], 4 * log(length(y)))
ob$update_partition()
ob$append_data(y[81:160], NULL)
ob$update_partition()
ob$get_partition()$changepoints
}
