% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-functions.dust.R
\name{dust}
\alias{dust}
\title{Perform DUST analysis to estimate selection coefficients on protein domains and identify those under significant positive selection.}
\usage{
dust(
  input.file.name,
  output.folder,
  output.prefix,
  steps = 1:7,
  N.sim = 1000,
  fp.cutoff.rate = 0.1,
  cnt.cutoff.rate = NULL,
  cnt.mis.cutoff = 6,
  cnt.non.cutoff = 6,
  cnt.both.cutoff = 6,
  rare.mut.cutoff = 3
)
}
\arguments{
\item{input.file.name:}{A VCF-formatted file to read SnpEff annotated somatic variants. This input file shall contain these fields: Tumor_Sample_Barcode, Chromosome, Start_Position, dbSNP_RS, Reference_Allele, Tumor_Seq_Allele2, FILTER, One_Consequence, Hugo_Symbol, Gene, Feature, ENSP, HGVSc, HGVSp_Short, Amino_acids, Codons, ENSP, RefSeq, Entrez_Gene_Id.}

\item{output.folder:}{the folder to write the temporary and final result files.}

\item{output.prefix:}{prefix used to name the temporary and final prediction files. Temporary and output files include prefix.outlier.txt, prefix.mut.filtered.txt, prefix.mut.region.txt, prefix.error.txt, prefix.exp_seq.by_type.region.txt, prefix.selection.region.txt, prefix.sig.region.txt, and a plot pdf file.}

\item{steps:}{A vector of integers indicating which functions to execute. 1-find.outliers(), 2-parse.aggregated.mut(), 3-annotate.by.cdd.region(), 4-calculate.expected.by.region, 5-compute.selection.by.region(), 6-find.sig.selection(), 7-plot.regions(). Default to 1:7.}

\item{N.sim:}{the number of simulations performed when calling the compute.selection.by.region() function. Default value is 1,000.}

\item{fp.cutoff.rate:}{false positive cutoff rate (default: 0.1).}

\item{cnt.non.cutoff:}{the minimum number of tumors harboring proteinitruncating mutations in a CDD region (default: 6).}

\item{cnt.mis.cutoff:}{the minimum number of tumors harboring missense or protein-truncating mutations in a CDD region (default: 6).}

\item{cnt.cutoff.rate:}{the minimum fraction of tumors harboring mutations in the CDD region (default: NULL). If set (range: 0 - 1), the minimum number of tumors will be calculated as round(cnt.cutoff.rate * number of samples in the prefix.mut.filtered.txt file. If this number is larger than cnt.both.cutoff, it will overwrite cnt.both.cutoff, cnt.mis.cutoff, and cnt.non.cutoff. If not set (NULL), cnt.mis.cutoff, cnt.non.cutoff, and cnt.both.cutoff must be set.}

\item{rare.mut.cutoff:}{the minimum number of missense or protein-truncating mutations (default: 3) in a gene for which to be considered as "rarely" mutated. For a rarely mutated gene, its gene name will be enclosed inside a parenthesis.}
}
\description{
Perform DUST analysis to estimate selection coefficients on protein domains and identify those under significant positive selection.
}
\examples{
gust(input.file.name='./examples/TCGA.CHOL.somatic.maf.gz', output.folder='./examples/', output.prefix='TCGA.CHOL');
}
