% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dust-package.R
\docType{package}
\name{dust}
\alias{dust}
\alias{dust-package}
\title{dust: Fast Multiple Change-Point Detection in Univariate Time Series}
\description{
The \pkg{dust} package implements DUST pruning for multiple change-point
detection. Its \code{dust.1D} interface supports eight one-parameter cost
models. The \code{dust.meanVar} interface detects changes in both the mean
and variance of a Gaussian series, using either one or two constraints.
Both interfaces offer scalar and optional Highway backends and online
objects for incremental analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/vrunge/dust}
  \item Report bugs at \url{https://github.com/vrunge/dust/issues}
}

}
\author{
\strong{Maintainer}: Vincent Runge \email{vincent.runge@univ-evry.fr}

Authors:
\itemize{
  \item Vincent Runge \email{vincent.runge@univ-evry.fr}
  \item Simon Querné \email{simon.querne@gmail.com}
}

Other contributors:
\itemize{
  \item Charles Truong \email{ctruong@ens-paris-saclay.fr} [contributor]
}

}
