% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dust-package.R
\docType{package}
\name{dust}
\alias{dust}
\alias{dust-package}
\title{dust: Fast Multiple Change-Point Detection}
\description{
Multiple change-point detection with the DUST pruning rule (DUality Simple Test).
\code{dust.1D} for univariate data (8 models), \code{dust.MD} for independent multivariate data
and \code{dust.meanVar} for changes in mean and variance. Each function has an object version
to add data step by step (\code{dust.object.1D}, \code{dust.object.MD}, \code{dust.object.meanVar}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/vrunge/dust}
  \item Report bugs at \url{https://github.com/vrunge/dust/issues}
}

}
\author{
\strong{Maintainer}: Vincent Runge \email{vincent.runge@univ-evry.fr}

Authors:
\itemize{
  \item Vincent Runge \email{vincent.runge@univ-evry.fr}
  \item Simon Querné \email{simon.querne@gmail.com}
  \item Félix Laplante \email{felixlaplante0@proton.me}
}

Other contributors:
\itemize{
  \item Charles Truong \email{ctruong@ens-paris-saclay.fr} [contributor]
}

}
