% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dust-package.R
\docType{package}
\name{dust}
\alias{dust}
\alias{dust-package}
\title{dust: Multiple Change-Point Detection in Multivariate Time Series}
\description{
The \pkg{dust} package implements algorithms for detecting multiple change
points in multivariate time series by minimizing a penalized likelihood.
It relies on optimal partitioning dynamic programming combined with the
DUality Simple Test (DUST) pruning method.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/vrunge/dust}
  \item Report bugs at \url{https://github.com/vrunge/dust/issues}
}

}
\author{
\strong{Maintainer}: Vincent Runge \email{vincent.runge@univ-evry.fr}

Authors:
\itemize{
  \item Simon Querné \email{simon.querne@gmail.com}
}

Other contributors:
\itemize{
  \item Charles Truong \email{ctruong@ens-paris-saclay.fr} [contributor]
}

}
