% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DUST_module_classes.R
\docType{class}
\name{DUST_module_classes}
\alias{DUST_module_classes}
\title{Internal S4 classes for DUST C++ modules}
\description{
Classes generated by the Rcpp module. Use \code{\link{dust.object.1D}}, \code{\link{dust.object.MD}} and \code{\link{dust.object.meanVar}}.
}
\seealso{
\code{\link{Rcpp_Detector-class}}, \code{\link{C++Object-class}}
}
\keyword{internal}
