% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DUST_module_classes.R
\docType{class}
\name{DUST_module_classes}
\alias{DUST_module_classes}
\alias{Rcpp_DUST_1D}
\alias{Rcpp_DUST_1D-class}
\alias{Rcpp_DUST_meanVar}
\alias{Rcpp_DUST_meanVar-class}
\alias{Rcpp_DUST_Reg}
\alias{Rcpp_DUST_Reg-class}
\alias{C++Object}
\alias{C++Object-class}
\title{Internal S4 classes for DUST C++ modules}
\description{
These S4 classes are generated by Rcpp modules and wrap the underlying
C++ implementations in C++. They are not intended to be used directly.
Instead, users should rely on the high-level interfaces:
\code{\link{dust.1D}}, \code{\link{dust.object.1D}},
\code{\link{dust.meanVar}} and \code{\link{dust.object.meanVar}}.
\code{\link{dust.Reg}} and \code{\link{dust.object.Reg}}.
}
\keyword{internal}
