% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{dust_package}
\alias{dust_package}
\title{Create dust model in package}
\usage{
dust_package(path)
}
\arguments{
\item{path}{Path to the package}
}
\value{
Nothing, this function is called for its side effects
}
\description{
Updates a dust model in a package. The user-provided code is
assumed to the in \code{inst/dust} as a series of C++ files; a file
\code{inst/dust/model.cpp} will be transformed into a file
\code{src/model.cpp}.
}
\details{
If your code provides a class \code{model} then dust will create C++
functions such as \code{dust_model_alloc} - if your code also includes
names such as this, compilation will fail due to duplicate
symbols.

We add "Rcpp attributes" to the created functions, and will run
\code{\link{compileAttributes}} on them once the generated code
has been created.
}
