% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{DUST_reg}
\alias{DUST_reg}
\title{MyModule: Exposing DUST_reg to R}
\description{
This module exposes the \code{DUST_reg} C++ class to R, allowing you to create
instances of \code{DUST_reg} and call its methods directly from R.
}
