% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rng.R
\name{dust_rng}
\alias{dust_rng}
\title{Dust Random Number Generator}
\value{
A \code{dust_rng} object, which can be used to drawn random
numbers from dust's distributions.
}
\description{
Create an object that can be used to generate random
numbers with the same RNG as dust uses internally.  This is
primarily meant for debugging and testing the underlying C++
rather than a source of random numbers from R.
}
\section{Running multiple streams, perhaps in parallel}{


The underlying random number generators are designed to work in
parallel, and with random access to parameters (see
\code{vignette("rng")} for more details).  However, this is usually
done within the context of running a model where each particle
sees its own stream of numbers.  We provide some support for
running random number generators in parallel, but any speed
gains from parallelisation are likely to be somewhat eroded by
the overhead of copying around a large number of random numbers.

All the random distribution functions support an argument
\code{n_threads} which controls the number of threads used.  This
argument will \emph{silently} have no effect if your installation
does not support OpenMP (see \link{dust_openmp_support}).

Parallelisation will be performed at the level of the stream,
where we draw \code{n} numbers from \emph{each} stream for a total of \code{n * n_streams} random numbers using \code{n_threads} threads to do this.
Setting \code{n_threads} to be higher than \code{n_streams} will therefore
have no effect. If running on somebody else's system (e.g., an
HPC, CRAN) you must respect the various environment variables
that control the maximum allowable number of threads; consider
using \link{dust_openmp_threads} to select a safe number.

With the exception of \code{random_real}, each random number
distribution accepts parameters; the interpretations of these
will depend on \code{n}, \code{n_streams} and their rank.
\itemize{
\item If a scalar then we will use the same parameter value for every draw
from every stream
\item If a vector with length \code{n} then we will draw \code{n} random
numbers per stream, and every stream will use the same parameter
value for every stream for each draw (but a different,
shared, parameter value for subsequent draws).
\item If a matrix is provided with one row and \code{n_streams}
columns then we use different parameters for each stream, but
the same parameter for each draw.
\item If a matrix is provided with \code{n} rows and \code{n_streams}
columns then we use a parameter value \verb{[i, j]} for the \code{i}th
draw on the \code{j}th stream.
}

The rules are slightly different for the \code{prob} argument to
\code{multinomial} as for that \code{prob} is a vector of values. As such
we shift all dimensions by one:
\itemize{
\item If a vector we use same \code{prob} every draw from every stream
and there are \code{length(prob)} possible outcomes.
\item If a matrix with \code{n} columns then vary over each draw (the
\code{i}th draw using vector \code{prob[, i]} but shared across all
streams. There are \code{nrow(prob)} possible outcomes.
\item If a 3d array is provided with 1 column and \code{n_streams}
"layers" (the third dimension) then we use then we use different
parameters for each stream, but the same parameter for each
draw.
\item If a 3d array is provided with \code{n} columns and \code{n_streams}
"layers" then we vary over both draws and streams so that with
use vector \code{prob[, i, j]} for the \code{i}th draw on the \code{j}th
stream.
}

The output will not differ based on the number of threads used,
only on the number of streams.
}

\examples{
rng <- dust::dust_rng$new(42)

# Shorthand for Uniform(0, 1)
rng$random_real(5)

# Shorthand for Normal(0, 1)
rng$random_normal(5)

# Uniform random numbers between min and max
rng$uniform(5, -2, 6)

# Normally distributed random numbers with mean and sd
rng$normal(5, 4, 2)

# Binomially distributed random numbers with size and prob
rng$binomial(5, 10, 0.3)

# Hypergeometric distributed random numbers with parameters n1, n2 and k
rng$hypergeometric(5, 6, 10, 4)

# Gamma distributed random numbers with parameters a and b
rng$gamma(5, 0.5, 2)

# Poisson distributed random numbers with mean lambda
rng$poisson(5, 2)

# Exponentially distributed random numbers with rate
rng$exponential(5, 2)

# Multinomial distributed random numbers with size and vector of
# probabiltiies prob
rng$multinomial(5, 10, c(0.1, 0.3, 0.5, 0.1))
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{info}}{Information about the generator (read-only)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-dust_rng-new}{\code{dust_rng$new()}}
\item \href{#method-dust_rng-size}{\code{dust_rng$size()}}
\item \href{#method-dust_rng-jump}{\code{dust_rng$jump()}}
\item \href{#method-dust_rng-long_jump}{\code{dust_rng$long_jump()}}
\item \href{#method-dust_rng-random_real}{\code{dust_rng$random_real()}}
\item \href{#method-dust_rng-random_normal}{\code{dust_rng$random_normal()}}
\item \href{#method-dust_rng-uniform}{\code{dust_rng$uniform()}}
\item \href{#method-dust_rng-normal}{\code{dust_rng$normal()}}
\item \href{#method-dust_rng-binomial}{\code{dust_rng$binomial()}}
\item \href{#method-dust_rng-hypergeometric}{\code{dust_rng$hypergeometric()}}
\item \href{#method-dust_rng-gamma}{\code{dust_rng$gamma()}}
\item \href{#method-dust_rng-poisson}{\code{dust_rng$poisson()}}
\item \href{#method-dust_rng-exponential}{\code{dust_rng$exponential()}}
\item \href{#method-dust_rng-multinomial}{\code{dust_rng$multinomial()}}
\item \href{#method-dust_rng-state}{\code{dust_rng$state()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dust_rng-new"></a>}}
\if{latex}{\out{\hypertarget{method-dust_rng-new}{}}}
\subsection{Method \code{new()}}{
Create a \code{dust_rng} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dust_rng$new(
  seed = NULL,
  n_streams = 1L,
  real_type = "double",
  deterministic = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{seed}}{The seed, as an integer, a raw vector or \code{NULL}.
If an integer we will create a suitable seed via the "splitmix64"
algorithm, if a raw vector it must the correct length (a multiple
of either 32 or 16 for \code{float = FALSE} or \code{float = TRUE}
respectively). If \code{NULL} then we create a seed using R's random
number generator.}

\item{\code{n_streams}}{The number of streams to use (see Details)}

\item{\code{real_type}}{The type of floating point number to use. Currently
only \code{float} and \code{double} are supported (with \code{double} being
the default). This will have no (or negligible) impact on speed,
but exists to test the low-precision generators.}

\item{\code{deterministic}}{Logical, indicating if we should use
"deterministic" mode where distributions return their
expectations and the state is never changed.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dust_rng-size"></a>}}
\if{latex}{\out{\hypertarget{method-dust_rng-size}{}}}
\subsection{Method \code{size()}}{
Number of streams available
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dust_rng$size()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dust_rng-jump"></a>}}
\if{latex}{\out{\hypertarget{method-dust_rng-jump}{}}}
\subsection{Method \code{jump()}}{
The jump function updates the random number state for
each stream by advancing it to a state equivalent to
2^128 numbers drawn from each stream.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dust_rng$jump()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dust_rng-long_jump"></a>}}
\if{latex}{\out{\hypertarget{method-dust_rng-long_jump}{}}}
\subsection{Method \code{long_jump()}}{
Longer than \verb{$jump}, the \verb{$long_jump} method is
equivalent to 2^192 numbers drawn from each stream.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dust_rng$long_jump()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dust_rng-random_real"></a>}}
\if{latex}{\out{\hypertarget{method-dust_rng-random_real}{}}}
\subsection{Method \code{random_real()}}{
Generate \code{n} numbers from a standard uniform distribution
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dust_rng$random_real(n, n_threads = 1L)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{Number of samples to draw (per stream)}

\item{\code{n_threads}}{Number of threads to use; see Details}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dust_rng-random_normal"></a>}}
\if{latex}{\out{\hypertarget{method-dust_rng-random_normal}{}}}
\subsection{Method \code{random_normal()}}{
Generate \code{n} numbers from a standard normal distribution
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dust_rng$random_normal(n, n_threads = 1L, algorithm = "box_muller")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{Number of samples to draw (per stream)}

\item{\code{n_threads}}{Number of threads to use; see Details}

\item{\code{algorithm}}{Name of the algorithm to use; currently \code{box_muller}
and \code{ziggurat} are supported, with the latter being considerably
faster.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dust_rng-uniform"></a>}}
\if{latex}{\out{\hypertarget{method-dust_rng-uniform}{}}}
\subsection{Method \code{uniform()}}{
Generate \code{n} numbers from a uniform distribution
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dust_rng$uniform(n, min, max, n_threads = 1L)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{Number of samples to draw (per stream)}

\item{\code{min}}{The minimum of the distribution (length 1 or n)}

\item{\code{max}}{The maximum of the distribution (length 1 or n)}

\item{\code{n_threads}}{Number of threads to use; see Details}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dust_rng-normal"></a>}}
\if{latex}{\out{\hypertarget{method-dust_rng-normal}{}}}
\subsection{Method \code{normal()}}{
Generate \code{n} numbers from a normal distribution
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dust_rng$normal(n, mean, sd, n_threads = 1L, algorithm = "box_muller")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{Number of samples to draw (per stream)}

\item{\code{mean}}{The mean of the distribution (length 1 or n)}

\item{\code{sd}}{The standard deviation of the distribution (length 1 or n)}

\item{\code{n_threads}}{Number of threads to use; see Details}

\item{\code{algorithm}}{Name of the algorithm to use; currently \code{box_muller}
and \code{ziggurat} are supported, with the latter being considerably
faster.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dust_rng-binomial"></a>}}
\if{latex}{\out{\hypertarget{method-dust_rng-binomial}{}}}
\subsection{Method \code{binomial()}}{
Generate \code{n} numbers from a binomial distribution
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dust_rng$binomial(n, size, prob, n_threads = 1L)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{Number of samples to draw (per stream)}

\item{\code{size}}{The number of trials (zero or more, length 1 or n)}

\item{\code{prob}}{The probability of success on each trial
(between 0 and 1, length 1 or n)}

\item{\code{n_threads}}{Number of threads to use; see Details}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dust_rng-hypergeometric"></a>}}
\if{latex}{\out{\hypertarget{method-dust_rng-hypergeometric}{}}}
\subsection{Method \code{hypergeometric()}}{
Generate \code{n} numbers from a hypergeometric distribution
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dust_rng$hypergeometric(n, n1, n2, k, n_threads = 1L)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{Number of samples to draw (per stream)}

\item{\code{n1}}{The number of white balls in the urn (called n in
R's \link{rhyper})}

\item{\code{n2}}{The number of black balls in the urn (called m in
R's \link{rhyper})}

\item{\code{k}}{The number of balls to draw}

\item{\code{n_threads}}{Number of threads to use; see Details}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dust_rng-gamma"></a>}}
\if{latex}{\out{\hypertarget{method-dust_rng-gamma}{}}}
\subsection{Method \code{gamma()}}{
Generate \code{n} numbers from a gamma distribution
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dust_rng$gamma(n, shape, scale, n_threads = 1L)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{Number of samples to draw (per stream)}

\item{\code{shape}}{Shape}

\item{\code{scale}}{Scale
'}

\item{\code{n_threads}}{Number of threads to use; see Details}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dust_rng-poisson"></a>}}
\if{latex}{\out{\hypertarget{method-dust_rng-poisson}{}}}
\subsection{Method \code{poisson()}}{
Generate \code{n} numbers from a Poisson distribution
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dust_rng$poisson(n, lambda, n_threads = 1L)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{Number of samples to draw (per stream)}

\item{\code{lambda}}{The mean (zero or more, length 1 or n)}

\item{\code{n_threads}}{Number of threads to use; see Details}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dust_rng-exponential"></a>}}
\if{latex}{\out{\hypertarget{method-dust_rng-exponential}{}}}
\subsection{Method \code{exponential()}}{
Generate \code{n} numbers from a exponential distribution
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dust_rng$exponential(n, rate, n_threads = 1L)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{Number of samples to draw (per stream)}

\item{\code{rate}}{The rate of the exponential}

\item{\code{n_threads}}{Number of threads to use; see Details}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dust_rng-multinomial"></a>}}
\if{latex}{\out{\hypertarget{method-dust_rng-multinomial}{}}}
\subsection{Method \code{multinomial()}}{
Generate \code{n} draws from a multinomial distribution.
In contrast with most of the distributions here, each draw is a
\emph{vector} with the same length as \code{prob}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dust_rng$multinomial(n, size, prob, n_threads = 1L)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{The number of samples to draw (per stream)}

\item{\code{size}}{The number of trials (zero or more, length 1 or n)}

\item{\code{prob}}{A vector of probabilities for the success of each
trial. This does not need to sum to 1 (though all elements
must be non-negative), in which case we interpret \code{prob} as
weights and normalise so that they equal 1 before sampling.}

\item{\code{n_threads}}{Number of threads to use; see Details}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dust_rng-state"></a>}}
\if{latex}{\out{\hypertarget{method-dust_rng-state}{}}}
\subsection{Method \code{state()}}{
Returns the state of the random number stream. This returns a
raw vector of length 32 * n_streams. It is primarily intended for
debugging as one cannot (yet) initialise a dust_rng object with this
state.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dust_rng$state()}\if{html}{\out{</div>}}
}

}
}
