% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rng.R
\name{dust_rng}
\alias{dust_rng}
\title{Dust Random Number Generator}
\description{
Create an object that can be used to generate random
numbers with the same RNG as dust uses internally.  This is
primarily meant for debugging and testing.
}
\examples{
rng <- dust::dust_rng$new(42)

# Shorthand for Uniform(0, 1)
rng$unif_rand(5)

# Shorthand for Normal(0, 1)
rng$norm_rand(5)

# Uniform random numbers between min and max
rng$runif(5, -2, 6)

# Normally distributed random numbers with mean and sd
rng$rnorm(5, 4, 2)

# Binomially distributed random numbers with size and prob
rng$rbinom(5, 10L, 0.3)

# Poisson distributed random numbers with mean lambda
rng$rpois(5, 2)
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{dust_rng$new()}}
\item \href{#method-size}{\code{dust_rng$size()}}
\item \href{#method-real_type}{\code{dust_rng$real_type()}}
\item \href{#method-jump}{\code{dust_rng$jump()}}
\item \href{#method-long_jump}{\code{dust_rng$long_jump()}}
\item \href{#method-unif_rand}{\code{dust_rng$unif_rand()}}
\item \href{#method-norm_rand}{\code{dust_rng$norm_rand()}}
\item \href{#method-runif}{\code{dust_rng$runif()}}
\item \href{#method-rnorm}{\code{dust_rng$rnorm()}}
\item \href{#method-rbinom}{\code{dust_rng$rbinom()}}
\item \href{#method-rpois}{\code{dust_rng$rpois()}}
\item \href{#method-rexp}{\code{dust_rng$rexp()}}
\item \href{#method-state}{\code{dust_rng$state()}}
\item \href{#method-set_deterministic}{\code{dust_rng$set_deterministic()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a \code{dust_rng} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dust_rng$new(seed, n_generators = 1L, real_type = "double")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{seed}}{The seed, as an integer or as a raw vector.}

\item{\code{n_generators}}{The number of generators to use. While this
function never runs in parallel, this is used to create a set of
interleaved independent generators as dust would use in a model.}

\item{\code{real_type}}{The type of floating point number to use. Currently
only \code{float} and \code{double} are supported (with \code{double} being
the default). This will have no (or negligible) impact on speed,
but exists to test the low-precision generators.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-size"></a>}}
\if{latex}{\out{\hypertarget{method-size}{}}}
\subsection{Method \code{size()}}{
Number of generators available
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dust_rng$size()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-real_type"></a>}}
\if{latex}{\out{\hypertarget{method-real_type}{}}}
\subsection{Method \code{real_type()}}{
Indicates the floating point type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dust_rng$real_type()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jump"></a>}}
\if{latex}{\out{\hypertarget{method-jump}{}}}
\subsection{Method \code{jump()}}{
The jump function for the generator, equivalent to
2^128 numbers drawn from the generator.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dust_rng$jump()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-long_jump"></a>}}
\if{latex}{\out{\hypertarget{method-long_jump}{}}}
\subsection{Method \code{long_jump()}}{
The \code{long_jump} function for the generator, equivalent
to 2^192 numbers drawn from the generator.
Generate \code{n} numbers from a standard uniform distribution
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dust_rng$long_jump()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-unif_rand"></a>}}
\if{latex}{\out{\hypertarget{method-unif_rand}{}}}
\subsection{Method \code{unif_rand()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dust_rng$unif_rand(n)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{Number of samples to draw
Generate \code{n} numbers from a standard normal distribution}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-norm_rand"></a>}}
\if{latex}{\out{\hypertarget{method-norm_rand}{}}}
\subsection{Method \code{norm_rand()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dust_rng$norm_rand(n)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{Number of samples to draw
Generate \code{n} numbers from a uniform distribution}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-runif"></a>}}
\if{latex}{\out{\hypertarget{method-runif}{}}}
\subsection{Method \code{runif()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dust_rng$runif(n, min, max)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{Number of samples to draw}

\item{\code{min}}{The minimum of the distribution (length 1 or n)}

\item{\code{max}}{The maximum of the distribution (length 1 or n)
Generate \code{n} numbers from a normal distribution}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rnorm"></a>}}
\if{latex}{\out{\hypertarget{method-rnorm}{}}}
\subsection{Method \code{rnorm()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dust_rng$rnorm(n, mean, sd)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{Number of samples to draw}

\item{\code{mean}}{The mean of the distribution (length 1 or n)}

\item{\code{sd}}{The standard deviation of the distribution (length 1 or n)
Generate \code{n} numbers from a binomial distribution}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rbinom"></a>}}
\if{latex}{\out{\hypertarget{method-rbinom}{}}}
\subsection{Method \code{rbinom()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dust_rng$rbinom(n, size, prob)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{Number of samples to draw}

\item{\code{size}}{The number of trials (zero or more, length 1 or n)}

\item{\code{prob}}{The probability of success on each trial
(between 0 and 1, length 1 or n)
Generate \code{n} numbers from a Poisson distribution}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rpois"></a>}}
\if{latex}{\out{\hypertarget{method-rpois}{}}}
\subsection{Method \code{rpois()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dust_rng$rpois(n, lambda)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{Number of samples to draw}

\item{\code{lambda}}{The mean (zero or more, length 1 or n)
Generate \code{n} numbers from a exponential distribution}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rexp"></a>}}
\if{latex}{\out{\hypertarget{method-rexp}{}}}
\subsection{Method \code{rexp()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dust_rng$rexp(n, rate)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{Number of samples to draw}

\item{\code{rate}}{The rate of the exponential}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-state"></a>}}
\if{latex}{\out{\hypertarget{method-state}{}}}
\subsection{Method \code{state()}}{
Returns the state of the random number generator. This returns a
raw vector of length 32 * n_generators. It is primarily intended for
debugging as one cannot (yet) initialise a dust_rng object with this
state.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dust_rng$state()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_deterministic"></a>}}
\if{latex}{\out{\hypertarget{method-set_deterministic}{}}}
\subsection{Method \code{set_deterministic()}}{
Toggle the RNG into/out of deterministic mode
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dust_rng$set_deterministic(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{Logical, \code{TRUE} if the RNG should run in deterministic
mode.}
}
\if{html}{\out{</div>}}
}
}
}
