% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation_cost.R
\name{segmentation_Cost_1D}
\alias{segmentation_Cost_1D}
\title{segmentation_Cost_1D}
\usage{
segmentation_Cost_1D(data, chpts, model = "gauss")
}
\arguments{
\item{data}{a numeric vector}

\item{chpts}{the change points (increasing, the last one is \code{length(data)})}

\item{model}{the model: \code{"gauss"} (default), \code{"poisson"}, \code{"exp"}, \code{"geom"}, \code{"bern"}, \code{"binom"}, \code{"negbin"}, \code{"variance"}}
}
\value{
the cost of the segmentation (without penalty)
}
\description{
Total cost of a segmentation (sum of the segment costs computed by \code{\link{Cost_1D}})
}
\details{
Costs are on the -2 log-likelihood scale, with segmentation-independent terms omitted. For \code{"gauss"}, \code{sum(data^2)} is added to obtain the residual sum of squares.
Binomial and negative binomial data have to be divided by the number of trials (or successes). The returned cost is the likelihood score divided by that size; divide the penalty by the same size.
}
\examples{
data <- dataGenerator_1D(chpts = c(300, 600, 900), parameters = c(0.6, 0.2, 0.4),
                         nbSuccess = 10, type = "negbin")
data <- data_normalization_1D(data, type = "negbin", size = 10)
segmentation_Cost_1D(data, c(300, 600, 900), model = "negbin")
segmentation_Cost_1D(data, c(150, 450, 900), model = "negbin")  # higher cost

}
\seealso{
\code{\link{Cost_1D}}
}
